v1
merged all three windows into a single class. removed the errors that presented themselves with this merge, and fixed some weird bugs that showed up in the program. the program is now fully functional, although crude visually and lacking in content. the game loop is completely programed, and now all that is left to do is add more events and update the visuals.